fix: 집사생활 댓글 controller 통합
API endpoint가 다르다고 착각하여 controller를 나누어 구현했으나
endpoint가 다르지 않았기에 boardController에 통합
