+++ server/catvillage/src/main/java/com/twentyfour_seven/catvillage/board/controller/BoardController.java
@@ -1,6 +1,7 @@
 package com.twentyfour_seven.catvillage.board.controller;
 
 import com.twentyfour_seven.catvillage.board.dto.*;
+import com.twentyfour_seven.catvillage.board.dto.comment.BoardCommentPatchDto;
 import com.twentyfour_seven.catvillage.board.dto.comment.BoardCommentPostDto;
 import com.twentyfour_seven.catvillage.board.dto.comment.BoardCommentPostResponseDto;
 import com.twentyfour_seven.catvillage.board.entity.Board;
@@ -149,5 +150,49 @@ public class BoardController {
         BoardComment boardComment = boardCommentMapper.boardCommentPostDtoToBoardComment(requestBody);
         BoardComment saveComment = boardCommentService.createBoardComment(user.getUsername(), boardComment);
         return new ResponseEntity<>(boardCommentMapper.boardCommentToBoardCommentPostResponseDto(saveComment), HttpStatus.CREATED);
+    }
+
+    @Operation(summary = "집사생활 댓글 조회",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "집사생활 댓글 조회 성공", content = @Content(schema = @Schema(implementation = BoardCommentPostResponseDto.class))),
+                    @ApiResponse(responseCode = "404", description = "존재하지 않는 집사생활 댓글")
+            })
+    @GetMapping("/{comments-id}")
+    public ResponseEntity<?> getBoardComment(@Valid @PathVariable("comments-id") Long boardCommentId) {
+        BoardComment findBoardComment = boardCommentService.findBoardComment(boardCommentId);
+        return new ResponseEntity<>(
+                boardCommentMapper.boardCommentToBoardCommentPostResponseDto(findBoardComment),
+                HttpStatus.OK);
+    }
+
+    @Operation(summary = "집사생활 댓글 수정",
+            responses = {
+                    @ApiResponse(responseCode = "200", description = "집사생활 댓글 수정 성공", content = @Content(schema = @Schema(implementation = BoardCommentPostResponseDto.class))),
+                    @ApiResponse(responseCode = "404", description = "존재하지 않는 집사생활 댓글\n존재하지 않는 유저"),
+                    @ApiResponse(responseCode = "405", description = "유저 정보 불일치")
+            })
+    @PatchMapping("/{comments-id}")
+    public ResponseEntity<?> patchBoardComment(@Valid @RequestBody BoardCommentPatchDto requestBody,
+                                               @Valid @Positive @PathVariable("comments-id") Long boardCommentId,
+                                               @AuthenticationPrincipal org.springframework.security.core.userdetails.User user) {
+        requestBody.setBoardCommentId(boardCommentId);
+        BoardComment boardComment = boardCommentMapper.boardCommentPatchDtoToBoardComment(requestBody);
+        BoardComment updateBoardComment = boardCommentService.updateBoardComment(user.getUsername(), boardComment);
+        return new ResponseEntity<>(
+                boardCommentMapper.boardCommentToBoardCommentPostResponseDto(updateBoardComment),
+                HttpStatus.OK);
+    }
+
+    @Operation(summary = "집사생활 댓글 삭제",
+            responses = {
+                    @ApiResponse(responseCode = "204", description = "집사생활 댓글 삭제 성공"),
+                    @ApiResponse(responseCode = "404", description = "존재하지 않는 집사생활 댓글\n존재하지 않는 유저"),
+                    @ApiResponse(responseCode = "405", description = "유저 정보 불일치")
+            })
+    @DeleteMapping("/{comments-id}")
+    public ResponseEntity<?> deleteBoardComment(@Valid @PathVariable("comments-id") Long boardCommentId,
+                                                @AuthenticationPrincipal org.springframework.security.core.userdetails.User user) {
+        boardCommentService.deleteBoardComment(user.getUsername(), boardCommentId);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
